refactor: separando funções genericas em outro arquivo global-controller
